Increase timeout to IMAP node.
Yahoo mail is unable to provide a response in under 3 seconds. Increasing
this limit to 10 seconds seems to solve the timeout issue.

I was unable to properly login to Yahoo but at least the timeout error
is not happening anymore.

Setting the node to work with Gmail worked fine although it is a
somewhat painful process as Gmail avoids the so called non secure apps.