Add SSL config to mysqldb connection
On postgres connection already is possible to configure ssl options
on connection.
with this commit, the intention is do the same to mysql.

To do that, was added 4 new env variables (DB_SSL_CA, DB_SSL_CERT,
DB_SSL_KEY, DB_SSL_REJECT_UNAUTHORIZED). These variables should work
to both database (mysql and postgres).

The postgres SSL variables were keeped as fallback, to not break who
is using the old way. In the future we should remove and just keep
theses variables introduced on this commit.