feat(core): Bundle UI in the published `n8n` package
Currently when someone installs n8n from npm, they also end up installing `n8n-editor-ui` and `n8n-design-system`.
With these changes, all UI code would get bundled up in the `cli` package, and none of the UI dependencies would get pulled in for a user installing n8n from npm.

The only downside of this approach is that the `n8n` package is larger now.
A better way to do this would be to remove all `dependencies` and `devDependencies` from `packages/editor-ui/package.json` before publishing that package. But, that would have to wait until the new release process is in place.