fix(core): Ensure that all non-lazy-loaded community nodes get post-processed correctly
If you install a community node with `polling: true`, activate a workflow with it fails with an error: `WorkflowActivationError: There was a problem activating the workflow: "Could not get parameter "pollTimes"!"`.

You can test this by installing `n8n-nodes-rss-feed-trigger`, creating a workflow with the `RSS Trigger` node, and then trying to activate it.
Activation will fail on `master`, but work as expected on this branch.