Display SSH private key as a multi-line text box, hide passphrase
In the latest n8n releases the private key credential input box displays as single line password input (as a result of #5871, a couple of months ago) rather than a multi-line text area. This is a problem as private keys are (always/almost always?) multiple lines of text. The result is that it is current impossible to enter a SSH key in the latest versions of n8n.

This PR reverts this change, while also making the passphrase a password field (which I suspect was the original intent).