fix(editor): Remove explicit parameter name scanning for code editors
The correct value to display in an editor is already passed in to `ParameterInput`. We do not need to scan for the correct parameter name again.
This also fixes the issue of us accidentally using `jsCode` as the parameter name in the Code node when the language is set to Python.