fix(workflowExecute): support pinning data for nodes with multiple outputs
The execution logic lacks a way to support pinning data for nodes with multiple outputs and forces pinned result data to have one element only. This commit fixes that so that n8n at least honors the pinned data length. Work will have to be conducted further to allow it from the UI as well.