Remove jest-injected args
From what I can tell, `0` is the index of the call. Jest tracks how many times the function `defaultReducer` has been called, and it provides the call index as an argument. And `[{ name: 'foo.bar', value: 'baz' }]` is the list of arguments that `defaultReducer` was called with. Jest records all the arguments passed to the spy function during each call and provides them in an array. It would seem these Jest-injected args are no longer compatible with the async reducer. Since these values are already being checked in the first two args, in the interest of time I decided to remove the Jest-injected arguments.