fix(core): fix pruning of non-finished executions
This fixes a bug in the pruning (soft-delete). The pruning was a bit too
aggressive, as it also pruned executions that weren't in an end state yet.
This only becomes an issue if there are long-running executions (e.g. workflow
with Wait node) or the prune parameters are set to keep only a tiny number of
executions.