feat(Gitlab Node): add support for pagination on getIssues
The GitLab API to get issues uses pagination:

https://docs.gitlab.com/ee/api/issues.html#list-issues

The default behaviour is to limit to 20 results per page.

The GitLab node does not respect this pagination.

I adjusted the existing parameters for `returnAll` and `limit`,
so the node now retrieves either all results of the search,
or limit to the given number.

Additionally, I extended the filter to support the parameter
`search`, searching for the parameter in the title and description
of issues.

Tests of the Gitlab node are not existing, so I did not add.

Disclaimer: I'm no developer. I followed the CONTRIBUTING guide. Please
let me know if additional changes are required to merge this feature.