fix(core): Account for activation error caused by follower `main` instance (#7735)
This PR accounts for the edge case where the user on a FE connected to a
follower `main` instance leads to an activation failure. The follower
performs the DB update, skips activation in `ActiveWorkflows`, instructs
leader to perform activation in `ActiveWorkflows`, leader fails to
activate and reverts the DB update and communicates activation error to
follower, which broadcasts the event to all connected FEs, which update
their stores and display the activation error.

To reproduce, create a workflow with a trigger with an invalid
credential and attempt to activate on both leader and follower to
trigger the activation error - behavior should be identical on both.