ci: Fix Docker build and use multi-step approach (no-changelog) (#7893)
Due to new dependencies released in `1.19`, the docker build is failing
due to missing build dependencie. To get around this, we split the build
into two phases. In the building phase, we install the build
dependencies required to compile new module bindings and copy the result
in the second phase.

Github issue / Community forum post (link here to close automatically):

---------

Signed-off-by: Oleg Ivaniv <me@olegivaniv.com>