fix(editor): Use web native <a> element in nav menus
Currently all navigation in the menus is done programmatically. This is not
accessible way to do navigation, as it prevents browser default behaviour,
like cmd/ctrl+click to open into new tab. Also screen readers don't give
any indication that the active element is an anchor.

This PR refactors component library's Menu and MenuItem to use either
vue-router RouterLink or <a> when the menu item is a navigation element.