exporting `refreshExpiringCookie` to unit test it
Additionally I had to wrangle the types a bit. I want to call
`await refreshExpiringCookie(...)` in the unit tests, but to do this
without TS complaining the type needs account for the function being
async. Right now adding the type to the variable binding hides that
fact.
My workaround for that is to let TS infer the type of
`refreshExpiringCookie`. That lead to TS not being able to infer the
types of `res` and `next`. To fix that I used `satisfied`.

Happy to apply other solutions to this like extending the RequestHandler
type and creating an `AsyncRequestHandler` or adding an `Async` type
helper and using that.

```ts
type Async<T extends (...args: unknown[]) => unknown> = (
	...args: Parameters<T>
) => Promise<ReturnType<T>>;
```