fix(core): Map NodeOperationErrors to BadRequestError in /resource-locator-results
UI calls the /dynamic-node-parameters/resource-locator-results endpoint to resolve
resources to show in the NDV resource picker. For example the user can select an
n8n workflow from a list in the n8n node. However, if the credentials are invalid or
missing, the resolution will throw a NodeOperationError. This results in HTTP 500
response from the endpoint. Since this is an expected error due to invalid
configuration, it makes more sense to respond with 400 Bad Request.