Send only data for the MFA option the user used
Before we were sending `mfaToken` and `mfaRecoveryCode` even though the user can only send one. It was working because we had an or condition in the backend. The changes on this PR change that condition, making it imposible to login with recovery codes. This change makes the FE to send only the field the user used making recovery codes work again.