fix(editor): Persist tag filter when clicking tag directly
The filters on the Workflows page are persisted when they are changed from
the filters dropdown. However they were not persisted if user clicked a tag
from the workflow card. This fixes that. That way user will see the same filters
when they come back to the workflows page from a workflow using back
button.