fix: Fix sending pin data twice causing payload too large errors
For manual executions the pin data is sent to the backend in the execute
request. At the moment it is included twice. Once in the `.pinData` and once
in the `.workflowData.pinData`. The data is identical. This can cause 413
payload too large errors for workflows for which it is still possible to pin
the data.

This change removes the `.pinData` parameter from the POST /:id/run
endpoint, requiring the pin data to be defined in the `.workflowData.pinData`
property. In the `IWorkflowExecutionDataProcess` interface the pin data
is kept in both `.pinData` and `.workflowData.pinData` properties for
backwards compatibility.