ci: Add GH workflow to notify about PR status changes
Add a workflow that calls a webhook for specific PR status changes.
This allows certain automations based on the event.

This could have also been implemented as a Github App or using a
personal access token, but decided to use a workflow for transparency.
The URL and access token are kept secret for security.