Feat(Text Classification Node): Change fallback option to a drop-down
The fallback option was not very clear in that it would drop items when
there would not be a clear category for a piece of text. Using a
drop-down option for the fallback, makes it explicit.

We also had plans for an 'always select closest match' option, but
scrapped that, because it was unreliable.