feat: Reintroduce collaboration feature
Collaboration feature was removed because it was missing authorization checks.
This adds checks that user can't send workflow opened or closed message, unless
they have access to that workflow. Also uses `cache` for storing the collaboration
state, so the feature works correctly with multiple mains.

The feature is currently only enabled for the new canvas, as that should come live
soon anyways.