build: Fix `cli` nodemon config
We use `nodemon` to watch for changing in files restart the n8n application
when the files change. At the moment we are watching the `commands` and
`src` directories as well as `index.ts`. However, `npm start` runs against
the built files from `dist` directory. That's why it makes more sense to actually
watch files in the `dist` directory.

More context:
https://n8nio.slack.com/archives/C03MZF137FV/p1725023150044619