fix(benchmark): Separate container data into own host folders
So different postgres instances have their own data instead of
reusing it from the previous run. Also fixes the permission issues
in the data directory. If we let docker create the mounted dirs,
they will be owned by root:root, and the containers won't have
access to them. Hence we create them ourselves and run the containers
using the current user and group