refactor(benchmark): Use `ssh` instead of `az ssh vm`
While the `az ssh vm` is more secure, it has multiple issues:
- The access token expires in 10 mins, causing test failures:
  `ERROR: AADSTS700024: Client assertion is not within its valid time range`
- It does not propagate the exit code, but instead always exits with 0
- It can't do `scp`, so we need to clone the repo. This makes local development more cumbersome
- It logs all kinds of unnecessary stuff into stdout

Hence this commits replaces `az ssh vm` with `ssh` and `scp`. Since the benchmark
environment is a throwaway env without any critical data, it's not the biggest deals.
We keep the private key hidden throughout the run anyways, so it's not leaked unless
someone gains access to the GH runner.