refactor(benchmark): Separate cloud env provisioning from running benchmarks
The access token in GH actions workflow expires very soon after creation. This
causes the cloud env destroying to fail because the benchmarks take some time
to run. This separates the provisioning and destroying of infra from running the
benchmarks. That way we can perform another login before destroying the env.