refactor(core): Include AI events in log streaming relay
This PR brings the AI events in line with all other log streaming events
in the relay, breaks a dependency cycle in `WorkflowExecuteAdditionalData`,
stops awaiting on log streaming during the AI nodes execution, and adds
tests for all AI log streaming events.

Notes:
- Removed `n8n.ai.output.parser.get.instructions` as it was unused.
- `cli` does not import `workflow` so I had to duplicate two small types.
Once we have a common types package this should no longer be an issue.

Follow-up to: #8526