ci: Fix branch & path ignores for e2e workflows
GH actions pull_request_review event does not filtering based on branch or paths.
This PR uses [paths-filter](https://github.com/dorny/paths-filter) action (or a fork of it to be precise)
to check if the chromatic or e2e-tests-pr workflows should run. Using a fork since the action
is missing an input from its yaml definition, but that feature has already been implemented.