feat: Implement auth for task runners
Authentication works in two steps:

1. `POST /rest/runners/auth` with `{ "token": "auth_token" }` to create a one-time use grant token
2. `/rest/runners/_ws?id=<runnerId>&token=<grant_token>` to initiate a task runner connection

The reason for this two phase auth mechanism is that in the next steps we can have a shim process
that takes care of the first step and starts the task runner only with the one-time use grant token.
That way the actual auth token is protected from anyone gaining access to the task runner.