feat: Add once for each item support for JS task runner
Make the JS task runner work with the 'runOnceForEachItem' execution mode. This
implementation is naive and not optimized, meaning each item is executed in
the task runner separately. We could optimize this so that this happens with
a single task, and the task runner iterates the items. In that case we do need
take into accoun the fact that the code could be defined as an expression.

This also makes the JS task runner output match 1:1 with the old sandbox
approach: errors are handled correctly (continue on fail), items are normalized
and validated, `item` and `items` aliases are provided in the execution context.