feat: Add support for $env in JS task runner
$env provides access to the instance's environment variables unless
it has been explicitly blocked. This commit adds functionality to pass
the environment variables to the JS task runner if needed.

I think ideally we wouldn't want to pass the entire env to the task
runner, but instead request them on-demand. However since we don't
run the JS code async, we can't do that for now. This shouldn't be
an issues tho since the env access is something you would want to
use only in a local installation.