Only load rows til key column row when appending rows to google sheet
This change reduces the amount of data that is been loaded when appending new rows to a google sheet. 

Before this change was the whole data set of the spreadsheet loaded, but only a few rows were needed to identity the columns.

The issue is that the provided "range" parameter of the execute function only contains the sheet name without any range annotations.

In case there would be any range annotations in the range, I extract the sheet name from it.