feat: Limit config file permissions on startup
Config file contains the instance encryption key, which should be treated as a secret.
We don't want it to be accessible by all users, so this PR changes the behaviour to
make it readable and writable only by the owner of the file (i.e. the same user which
the n8n application is running as). This behaviour is similar to what `ssh` does with
private keys, i.e. it refuses to use the key unless it has limited permissions.