fix: Provide a better error message when task runner disconnects
When a task runner executing Code Node crashes, it's nice to get a proper
error message. This adds detection of Out of Memory (OOM) for task runners
running "internally" (i.e. as a child process).

Since the failure of the task (detected when task runner disconnects)
and crash of the child process happen concurrently, we need to sync
them. I used a sliding window to get the exit reason of the process
when the runner disconnects. This way it doesn't matter if we get
the exit reason from the process before, during or after the disconnect
happens, as long as it happens within the defined time window (500ms).

Other option would have been to use the runnerId, and get the exit
reason based on the ID. I didn't go this route mainly because it
would have required more changes to runner itself (i.e. instead of
generating the ID on the runner, we generate it outside and match
the runner in the task runner server and in process with the id).
This would be more reliable as with the sliding window there might
be incorrect reasons if the runner is e.g. in a crash loop.