feat: Only send needed data to task runner
When executing JS code in the Code Node, the task runner has currently
been fetching the entire workflow execution context data. This can
be a lot of data, and can cause OOMs on large workflows. Most often
Code Node is used in such a way that it only uses the input data and
maaaaybe some other node's data. Hence sending all the data is an
overkill.

This PR changes the behaviour to send only the needed data. This is
implemented by running a static analysis of the code and identifying
which built-in variables are accessed within the code. Based on that
analysis we only send the needed data.

There can be corner cases where we can't statically analyse for
example which nodes' data is needed. This can happen for example
when a variable is used as parameter to $() function. In these
cases we send all the data.

Based on some naive measurements, the message size is reduced
something between 30-70%.

Next steps:
After this change there is room for optimization. For example,
there is still duplication in the message send to the task runner.
The input data of the Code Node is located in multiple parts of
the message, which all end up being separate objects when
deserialized in the task runner.