fix(core): Increase task runner offer validity and introduce randomness
The one second expire time we had before is quite tight if the runner
or the n8n instance is under load. This PR increases it to 5s, so we
don't have tasks expiring too eagerly. Also introduces a bit of
randomness into the validity, so the offers don't expire all at
the same time.