fix(Postgres) Allow users to wrap strings with `$$`
In utils, `pgp.helpers.concat`, calls `pg.format` under the hood.
`pg.format` interprets any combination of a dollar sign with a number
(eg: `$1`, `$2`) as a query parameter regardless of a prepended dollar
sign (ie: `$$1` is interpreted as expecting a query parameter in
position 1). We need to allow for partial replacement, which involves
passing in an options object.