fix(core): Ensure tasks timeout even if they don't receive settings
If the n8n instance happens to crash on a specific time, task runner task
might not receive a response to e.g. settings or data request. In cases like
this the task runner was left hanging forever. This PR makes sure the tasks
get aborted correctly.

Also refactors the task execution lifecycle to be explicit which states the task
might have and how different events are handled in different states.