fix(core): Fix usage of external libs in task runner
After freezing built-ins to prevent prototype pollution, certain libraries like
moment stopped working. This is because the overwrite certain functions
on their own prototype that are defined in the Object's prototype. Usually
this is `toString`. This fixes that by importing the library before freezing,
so the allowed external libraries can set up their internal classes.