refactor(core): Use IWorkflowBase over WorkflowEntity in most places
In general, it's better to depend on abstractions over concrete implementations.
Also it's better to depend on a minimal required interface over a fat one.
Depending on `WorkflowEntity` ties much of our code to our data access layer,
and makes the code rely on much more than it needs to. This makes it more
difficult to introduce changes in the future to the data layer.