feat(Vector Store): Make vector store retrievers accept either tool or vector store inputs (wip)
Change vector store nodes to accept both AiTool and AiVectorStore outputs in retrieve mode,
and drop the retrieve-as-tool mode from being an option. "retrieve" mode now has a sub "retrieveMode"
selection which checks based on its output connections whether to act in tool or vector store connection mode
by utilizing a dynamic default connection on a disabled field that has an expression that evaluates to
either "retrieve" or "retrieve-as-tool".

Unfortunately currently this disabled internal field cannot be hidden, or else the value won't be evaluated.
To actually support this we would need to build some kind of support for hidden internal fields that would
still get their expressions evaluated...

This PoC was built for a spike ADO-3447.