remove code that never worked correctly
This code registers two handlers on the proxy, one for error and one for
connection, then resolves the promise.
Then after the promise is resolved we try to connect to the proxy, which
triggers one of the two handlers or both.
If it triggers the error handler we try to reject the resolved promise,
which will do nothing, since promises can only settle once.

So all of the error handling code was never called.