refactor(core): Simplify how webhook response is generated in `lastNode` mode
WebhookHelper's `executeWebhook` function does a lot. This PR attempts make it a bit more manageable by extracting how the response payload is created. It also centralizes the actual
sending of the response to `WebhookRequestHandler` based on what `executeWebhook`
resolves using the `responseCallback`.

It is possible this change has unforeseen consequences as there can be subtle differences in
the implementation in different modes and nodes.