fix(core): Use CSP header to sandbox html webhooks instead of iframe
In #16978 we fixed a security issue by sandboxing HTML responses from webhooks. The sandboxing
isolates the HTML page's origin. We used an `iframe` as a mechanism to separate the page's origin.
Probably due to an oversight, we weren't aware that you can also use CSP headers to do the same
kind of sandboxing.

This PR changes the sandboxing mechanism to use CSP header instead of an iframe. This is much less
intrusive, as we don't have to touch the response body. It also simplifies the implementation quite a bit.