SQL AG Multi Subnet Option
When SQL availability clusters have cluster members in separate subnets the connection string requires the   multiSubnetFailover option to be set to true. If this option is not set the connection string will  fail if the DNS lookups return one of the passive nodes.  The change required to add this functionality was a new option in the SQL pool code and matching Boolean option on credential definition.