fix(core): Set more reasonable defaults for scaling mode queue lock
In scaling mode, `bull` uses a lock to reserve the job for a specific worker. The lock has
an expiration time that is used to detect worker crashes and stalls. The default values
for the lock duration are 30s and by default it's renewed every 15s (in practice longer
due to the renewal is scheduled only after the new lock has been acquired).

It has been noticed that under heavy load workers might not be able to renew the lock
during that time, which causes `Queue errored` and `Missing lock` errors. Depending
on the `QUEUE_WORKER_MAX_STALLED_COUNT` value, this might also cause `bull`
to retry the workflow execution.

This PR increases the defaults from 30s and 15s to 60s and 10s. This gives workers more
leeway to renew the lock during heavy load.

It would also make sense to change the default of `QUEUE_WORKER_MAX_STALLED_COUNT`
to 0, or remove it altogether, but we probably have to do that in v2.