fix(core): Add a signature to webhook's resume url
Generate and include a signature for the workflow's resume URL. This is to
secure the URL from sniffing. Also version the `IRunExecutionData` interface,
so we know if the execution was created with an n8n version that includes the
signature in the URL. That way we can keep this backwards compatible.