fix(editor): Don't dismiss what's new on workflow execution
The "whats new" toast is closed by the "clearAllStickyNotifications" when clicking "execute workflow".
But since the onClose handler of the notification also sets the localstorage prop
which result in the toast no longer being shown, we auto-dismiss this toast even though the user didn't actively dismiss it.

With this change, the only way to dismiss a "whats new" toast, is for the user to click on it and have the modal opened with the "new stuff" article.