refactor(core): Remove types for request parameter that's not implemented
While the ExecutionRepository.findSingleExecution method supports an "unflattenData" option, the implementation of ExecutionService.findOne uses the ExecutionRepository.findIfShared method,
which hardcodes the "unflattenData" option to false.
The ExecutionService never passes the "unflattenResponse" option described in the removed "GetOne" type
to the ExecutionRepository, so it can't have any effect.
Additionally, the use of true and false strings for a boolean option is not ideal.