Fix source control connection being setup twice
While the form element is the correct semantic element to use here, we'd need to first rearrange the elements on the page a bit.

Because in the state where the instance is not connected, there is just one button on the page "Connect".
This would be the submit button of the form.

Once the instance is connected however, there's an additional "Disconnect git" button right next to the repository url.
This one will take on the role of the "submit button" inside the form element, even though it is not submitting the form, but performing an action.

So before we introduce the form element, we need to reposition the disconnect button, e.g. by moving it to bottom of the inputs, outside of the form.