Mark workflowStatus prop as optional
While the modal behaves more reliably when workflowStatus is defined,
since it will not fetch the remote version of a workflow that hasn't been pushed yet,
we're opting to not adjust the "shouldOpen" logic inside useWorkflowDiffRouting::handleDiffModal,
to be false when workflowStatus is undefined.

This way, our code is just a little more flexible in case the workflowStatus
ends up not being present in the URL's query parameters for some reason.